Migrate matrix-scalar multiplication tests
