+++ src/unittests/f32_interval.spec.ts
@@ -17,7 +17,6 @@ import {
   correctlyRoundedInterval,
   faceForwardIntervals,
   modfInterval,
-  multiplicationMatrixScalarInterval,
   multiplicationMatrixVectorInterval,
   multiplicationVectorMatrixInterval,
   refractInterval,
@@ -2031,153 +2030,6 @@ g.test('modfInterval')
       `modfInterval([${t.params.input}) returned { fract: [${got.fract}], whole: [${got.whole}] }. Expected { fract: [${expected.fract}], whole: [${expected.whole}] }`
     );
   });
-
-interface MatrixScalarToMatrixCase {
-  matrix: number[][];
-  scalar: number;
-  expected: (number | IntervalBounds)[][];
-}
-
-g.test('multiplicationMatrixScalarInterval')
-  .paramsSubcasesOnly<MatrixScalarToMatrixCase>([
-    // Only testing that different shapes of matrices are handled correctly
-    // here, to reduce test duplication.
-    // multiplicationMatrixScalarInterval uses MultiplicationIntervalOp for calculating intervals,
-    // so the testing for multiplcationInterval covers the actual interval
-    // calculations.
-    {
-      matrix: [
-        [1, 2],
-        [3, 4],
-      ],
-      scalar: 10,
-      expected: [
-        [10, 20],
-        [30, 40],
-      ],
-    },
-    {
-      matrix: [
-        [1, 2],
-        [3, 4],
-        [5, 6],
-      ],
-      scalar: 10,
-      expected: [
-        [10, 20],
-        [30, 40],
-        [50, 60],
-      ],
-    },
-    {
-      matrix: [
-        [1, 2],
-        [3, 4],
-        [5, 6],
-        [7, 8],
-      ],
-      scalar: 10,
-      expected: [
-        [10, 20],
-        [30, 40],
-        [50, 60],
-        [70, 80],
-      ],
-    },
-    {
-      matrix: [
-        [1, 2, 3],
-        [4, 5, 6],
-      ],
-      scalar: 10,
-      expected: [
-        [10, 20, 30],
-        [40, 50, 60],
-      ],
-    },
-    {
-      matrix: [
-        [1, 2, 3],
-        [4, 5, 6],
-        [7, 8, 9],
-      ],
-      scalar: 10,
-      expected: [
-        [10, 20, 30],
-        [40, 50, 60],
-        [70, 80, 90],
-      ],
-    },
-    {
-      matrix: [
-        [1, 2, 3],
-        [4, 5, 6],
-        [7, 8, 9],
-        [10, 11, 12],
-      ],
-      scalar: 10,
-      expected: [
-        [10, 20, 30],
-        [40, 50, 60],
-        [70, 80, 90],
-        [100, 110, 120],
-      ],
-    },
-    {
-      matrix: [
-        [1, 2, 3, 4],
-        [5, 6, 7, 8],
-      ],
-      scalar: 10,
-      expected: [
-        [10, 20, 30, 40],
-        [50, 60, 70, 80],
-      ],
-    },
-    {
-      matrix: [
-        [1, 2, 3, 4],
-        [5, 6, 7, 8],
-        [9, 10, 11, 12],
-      ],
-      scalar: 10,
-      expected: [
-        [10, 20, 30, 40],
-        [50, 60, 70, 80],
-        [90, 100, 110, 120],
-      ],
-    },
-    {
-      matrix: [
-        [1, 2, 3, 4],
-        [5, 6, 7, 8],
-        [9, 10, 11, 12],
-        [13, 14, 15, 16],
-      ],
-      scalar: 10,
-      expected: [
-        [10, 20, 30, 40],
-        [50, 60, 70, 80],
-        [90, 100, 110, 120],
-        [130, 140, 150, 160],
-      ],
-    },
-  ])
-  .fn(t => {
-    const matrix = t.params.matrix;
-    const scalar = t.params.scalar;
-    const expected = toF32Matrix(t.params.expected);
-    const got = multiplicationMatrixScalarInterval(matrix, scalar);
-    t.expect(
-      objectEquals(expected, got),
-      `multiplicationMatrixScalarInterval([${JSON.stringify(
-        matrix
-      )}], ${scalar}) returned '[${JSON.stringify(got)}]'. Expected '[${JSON.stringify(expected)}]'`
-    );
-  });
-
-// There are no explicit tests for multiplicationScalarMatrixInterval, since it
-// is just a pass-through to multiplicationMatrixScalarInterval
 
 interface MatrixVectorToVectorCase {
   matrix: number[][];
